Move remaining pulp_file test utils
[noissue]
