+++ pulpcore/tests/functional/api/using_plugin/test_content_access.py
@@ -6,7 +6,7 @@ from pulpcore.client.pulp_file import (
     RepositorySyncURL,
 )
 
-from pulp_file.tests.functional.utils import (
+from pulpcore.tests.functional.utils import (
     download_file,
 )
